Add ICACHE_RAM_ATTR to fix Arduino specific linking defaults
The SDK puts everything in the ICACHE memory by default unless marked by
ICACHE_FLASH_ATTR, which moves the code to the faultable IROM section.

Arduino inverses the default and adds the ICACHE_RAM_ATTR for code that is
not allowed to fault (i.e. eventulally has to be fetched from the SPI flash),
e.g. ISR code.
Fixes issue #11.